Coding style fixes
