+++ tslint-rules/src/utils/path.ts
@@ -1,0 +1,14 @@
+import * as Path from 'path';
+
+export function removeQuotes(value: string): string {
+  let groups = /^(['"])(.*)\1$/.exec(value);
+  return groups ? groups[2] : '';
+}
+
+export function removeModuleFileExtension(fileName: string): string {
+  return fileName.replace(/\.(?:(?:js|ts)x?|d\.ts)?$/i, '');
+}
+
+export function getBaseNameWithoutExtension(fileName: string) {
+  return Path.basename(fileName, Path.extname(fileName));
+}
